functional motivate button
